feat(api): implement core backend MVP for UdyogSaarthi

+++ backend/app/schemas/directory.py
@@ -5,7 +5,7 @@ class NearbyProfile(BaseModel):
     id: str
     name: str
     category: str
-    distance_m: int
+    distance_m: float
     lat: float
     lon: float
